audio-to-image-app/requirements.txt
streamlit==1.33.0
groq==0.11.0
python-dotenv==1.0.1
requests==2.31.0
pydub==0.25.1
mistralai==1.0.1
Pillow==10.2.0
clipdrop-sdk==1.0.0